\name{GENESIS-defunct}
\alias{GENESIS-defunct}
\alias{assocTestSeq}
\alias{assocTestSeqWindow}
\title{Defunct functions in package \pkg{GENESIS}}

\description{
  These functions are defunct and no longer available.
}

\details{

  The following functions are defunct; use the replacement indicated below:
  \itemize{

    \item{assocTestSeq: \code{\link{assocTestAggregate}}}
    \item{assocTestSeqWindow: \code{\link{assocTestAggregate}}}

  }
}