\name{GENESIS-deprecated}
\alias{GENESIS-deprecated}
\title{Deprecated functions in package \pkg{GENESIS}}

\description{
  These functions are provided for compatibility with older versions
  of \pkg{GENESIS} only, and will be defunct at the next release.
}

\details{

  The following functions are deprecated and will be made defunct; use
  the replacement indicated below:
  \itemize{

    \item{admixMapMM: \code{\link{admixMap}}}
    \item{assocTestMM: \code{\link{assocTestSingle}}}
    \item{fitNullMM: \code{\link{fitNullModel}}}
    \item{fitNullReg: \code{\link{fitNullModel}}}
    \item{king2mat: \code{\link{kingToMatrix}}}
    \item{pcrelateReadInbreed: \code{\link{pcrelate}}}
    \item{pcrelateReadKinship: \code{\link{pcrelate}}}

  }
}