\name{GENESIS-package}
\alias{GENESIS-package}
\alias{GENESIS}
\docType{package}
\title{
\packageTitle{GENESIS}
}
\description{
\packageDescription{GENESIS}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{GENESIS}
\packageIndices{GENESIS}
GENetic Effect-Size distribution Inference from Summary-level data
}
\author{
\packageAuthor{GENESIS}

Maintainer: \packageMaintainer{GENESIS}
}
\references{
Zhang, Yan, et al. "Estimation of complex effect-size distributions using summary-level statistics from genome-wide association studies across 32 complex traits" Nature Genetics (2018).
}
\keyword{effect size distribution, GWAS, mixture model }
