\name{HapMap_ASW_MXL_KINGmat}
\alias{HapMap_ASW_MXL_KINGmat}
\docType{data}
\title{Matrix of Pairwise Kinship Coefficient Estimates for the combined HapMap ASW and MXL Sample found with the KING-robust estimator from the KING software.}
\description{KING-robust kinship coefficient estimates for the combined HapMap African Americans in the Southwest U.S. (ASW) and Mexican Americans in Los Angeles (MXL) samples.}
\value{A matrix of pairwise kinship coefficient estimates as calculated with KING-robust for the combined HapMap African Americans in the Southwest U.S. (ASW) and Mexican Americans in Los Angeles (MXL) samples.}
\usage{data(HapMap_ASW_MXL_KINGmat)}
\format{
  The format is:
 num [1:173, 1:173] 0 0.00157 -0.00417 0.00209 0.00172 ...
}
\source{
	http://hapmap.ncbi.nlm.nih.gov/
}
\references{
	International HapMap 3 Consortium. (2010). Integrating common and rare genetic variation in diverse human populations. Nature, 467(7311), 52-58.
}
\keyword{datasets}
