\name{postmarginal}
\alias{postmarginal}
\title{
Posterior Inference Function of Marginal Effect Size
}
\description{
This function allows to estimate posterior probability with a null marginal effect size and posterior mean of marginal effect size. 
}
\usage{
postmarginal(summarydata, est, c0=10, LDcutoff=0.1, LDwindow=1, filter=FALSE)
}
\arguments{
 \item{\code{summarydata}}{either summay-level GWAS data, containing 3 columns: SNP (SNP rsID),  Z (GWAS test z-statistic), N (GWAS study sample size which can be different for different SNPs); or a dataframe generated by preprocessing() function  containing the filtered summary GWAS data  after merged with the LD information, with 10 columns
("SNP", "Z", "N", "betahat", "varbetahat", "Nstar", "TaggingSNPs", "LD.score.correct", "CHR", "BP").   
}
  \item{\code{est}}{parameter estimates by fitting either 2-component model, i.e., (pic, sigmasq, a); or 3-component model, i.e., (pic, p1, sigmasq1, sigmasq2, a).}
    \item{\code{c0}}{an assumed maximum number of underlying susceptibility SNPs tagged by any individual GWAS marker. By default, c0 is set at 10.}
 \item{\code{LDcutoff}}{a number from (0.05, 0.1, 0.2); indicating LD score is calculated based on the particular r^2 cutoff. By default, it is 0.1.}
   \item{\code{LDwindow}}{a number from (0.5, 1, 2); indicating LD score is calculated based on the particular window size (MB). By default, it is 1 MB.}
  \item{\code{filter}}{logical; if TRUE, the input summary data will be filtered.}
}
\details{
This function computes the posterior probability for each GWAS maker to have  a zro marginal effect, as well as the best prediction of marginal effect size given the GWAS summary-level statistics and the fitted effect size distribution. See Supplementary Note Section 4  in  Zhang et. al 2018.
}
\value{
A list  with components: 
\item{\code{snp}}{SNP list of the marginal effect size.}
\item{\code{prob.beta0}}{posterior probability for the marker to have a zero marginal effect.}
\item{\code{posteriormean.marginal}}{posterior mean (best prediction) of marginal effect size.}
}
\references{
Zhang, Yan, et al. "Estimation of complex effect-size distributions using summary-level statistics from genome-wide association studies across 32 complex traits" Nature Genetics (2018).
}
\author{
Yan Zhang <yandorazhang@gmail.com>
}
\examples{
data(heightGWAS)
post <- postmarginal(heightGWAS,est=c(8.899809e-03, 9.476025e-02, 1.458650e-04, 2.227118e-05, 1.567643e-06), c0=10, LDcutoff=0.1, LDwindow=1, filter=FALSE)
head(post$snp)
summary(post$prob.beta0)
summary(post$posteriormean.marginal)
}
