\name{qqplotdata.simu}
\alias{qqplotdata.simu}
\title{
Function to Generate Dataframe needed for QQ Plot
}
\description{
This function allows to generate expected p-values under the fitted model through simulations, hence then get the data needed to create QQ plot comparing observed distributions of association statistics against those expected under the fitted models. 
}
\usage{
qqplotdata.simu(summarydata, est, c0=10, qqplotCI.coverage=0.8, qqplot.nsim=100, LDcutoff=0.1, LDwindow=1, filter=F,cores=1,seeds=123)
}
\arguments{
 \item{\code{summarydata}}{either summay-level GWAS data, containing 3 columns: SNP (SNP rsID),  Z (GWAS test z-statistic), N (GWAS study sample size which can be different for different SNPs); or a dataframe generated by preprocessing() function  containing the filtered summary GWAS data  after merged with the LD information, with 10 columns
("SNP", "Z", "N", "betahat", "varbetahat", "Nstar", "TaggingSNPs", "LD.score.correct", "CHR", "BP").   
}
  \item{\code{est}}{parameter estimates by fitting either 2-component model, i.e., (pic, sigmasq, a); or 3-component model, i.e., (pic, p1, sigmasq1, sigmasq2, a).}
  \item{\code{c0}}{an assumed maximum number of underlying susceptibility SNPs tagged by any individual GWAS marker. By default, c0 is set at 10.}
\item{\code{qqplotCI.coverage}}{the coverage rate of confidence band in the QQ plot. By default, it is 0.8.}
  \item{\code{qqplot.nsim}}{the total number of simulations to generate the expected p-values to get the confidence band in QQ plot. By default, it is 100.}
 \item{\code{LDcutoff}}{a number from (0.05, 0.1, 0.2); indicating LD score is calculated based on the particular r^2 cutoff. By default, it is 0.1.}
   \item{\code{LDwindow}}{a number from (0.5, 1, 2); indicating LD score is calculated based on the particular window size (MB). By default, it is 1 MB.}
  \item{\code{filter}}{logical; if TRUE, the input summary data will be filtered.}
  \item{\code{cores}}{number of CPU threads in parallel computing; by default, it is 1.}
  \item{\code{seeds}}{numeric, random seeds used in simulation; by default, it is 123.}
}
\details{
This function allows to generate expected p-values under the fitted model through simulations, hence then get the dataframe to generate qqplot. The details are as follows: (i) generate marginal effect size from mixture normal distribution (eq.(6) in Supplementary Note in Zhang et. al 2018)  with parameters  plugged-in by their estimated values from the fitted model; (ii) add error terms following N(0, R) distribution to get the estimated marginal effect sizes (betahat); (iii) transform the simulated  "betahat" to underlying p-values. (see Supplementary Note Section 5.1 in Zhang et. al 2018)
}
\value{
A list containing
\item{\code{QQdata}}{a dataframe needed to generate the QQ plot, i.e., a dataframe containing 4 columns, "log10(observed pvalues)", "expected log10(fitted pvalues)", "lower bound of log10(fitted pvalues)","upper bound of log10(fitted pvalues)". }
\item{\code{observedlambda}}{lambda genomic control (GC) factor for observed summary GWAS statistics.}
\item{\code{meanEXPlambda}}{expected lambda GC factor for simulated summary GWAS statistics under the fitted model.}
\item{\code{lowEXPlambda}}{qqplotCI.coverage CI lower bound of lambda GC factor for simulated summary GWAS statistics under the fitted model.}
\item{\code{highEXPlambda}}{qqplotCI.coverage CI  upper bound of lambda GC factor for simulated summary GWAS statistics under the fitted model.}
}
\references{
Zhang, Yan, et al. "Estimation of complex effect-size distributions using summary-level statistics from genome-wide association studies across 32 complex traits" Nature Genetics (2018).
}
\author{
Yan Zhang <yandorazhang@gmail.com>
}
\seealso{
\code{\link{qqplot.plot}}
}
\examples{
data(heightGWAS)
qqplotdata <- qqplotdata.simu(heightGWAS, est=c(9.583307e-03,8.562964e-02,1.487684e-04,2.086576e-05,1.498790e-06))
qqplot.plot(qqplotdata)
}
