\name{w_hm3.noMHC.snplist}
\alias{w_hm3.noMHC.snplist}
\docType{data}
\title{SNP List of HapMap 3 SNPs after Extracting SNPs within MHC Region}
\description{SNP list of HapMap 3 SNPs after extracting SNPs within the major histocompatibility complex (MHC) region}
\usage{data("w_hm3.noMHC.snplist")}
\format{
  A matrix with 1215001 rows and 3 columns (SNP, A1, A2). 
}
\examples{data("w_hm3.noMHC.snplist")}
\keyword{datasets}
