\name{GENESIS-defunct}
\alias{GENESIS-defunct}
\alias{admixMapMM}
\alias{assocTestMM}
\alias{assocTestSeq}
\alias{assocTestSeqWindow}
\alias{fitNullMM}
\alias{fitNullReg}
\alias{king2mat}
\alias{pcrelateMakeGRM}
\alias{pcrelateReadInbreed}
\alias{pcrelateReadKinship}
\alias{pcrelate,GenotypeData-method}
\alias{pcrelate,SeqVarData-method}
\title{Defunct functions in package \pkg{GENESIS}}

\description{
  These functions are defunct and no longer available.
}

\details{

  The following functions are defunct; use the replacement indicated below:
  \itemize{

    \item{admixMapMM: \code{\link{admixMap}}}
    \item{assocTestMM: \code{\link{assocTestSingle}}}
    \item{assocTestSeq: \code{\link{assocTestAggregate}}}
    \item{assocTestSeqWindow: \code{\link{assocTestAggregate}}}
    \item{fitNullMM: \code{\link{fitNullModel}}}
    \item{fitNullReg: \code{\link{fitNullModel}}}
    \item{king2mat: \code{\link{kingToMatrix}}}
    \item{pcrelate,GenotypeData-method: \code{\link{pcrelate,GenotypeIterator-method}}}
    \item{pcrelate,SeqVarData-method: \code{\link{pcrelate,SeqVarIterator-method}}}
    \item{pcrelateMakeGRM: \code{\link{pcrelateToMatrix}}}
    \item{pcrelateReadInbreed: \code{\link{pcrelate}}}
    \item{pcrelateReadKinship: \code{\link{pcrelate}}}

  }
}